backend/requirements.txt
asgiref==3.2.7
autopep8==1.5.3
Django==3.0.7
django-cors-headers==3.3.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
importlib-metadata==1.6.1
Markdown==3.2.2
Pillow==7.1.2
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
zipp==3.1.0
